requirements.txt
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==7.1.2
